Reorganizes test files
