+++ lib/breakpoint-match.test.js
@@ -1,4 +1,4 @@
-const breakpointMatch = require('../lib/breakpoint-match');
+const breakpointMatch = require('./breakpoint-match');
 
 describe('Find a breakpoint match: px', () => {
   // Parsed options.
