+++ lib/normalize-options.test.js
@@ -1,4 +1,4 @@
-const { normalizeOptions, handleBreakpointConfigs } = require('../lib/normalize-options');
+const { normalizeOptions, handleBreakpointConfigs } = require('./normalize-options');
 
 /**
  * Test expected option values after normalization.
