+++ lib/parse-options.test.js
@@ -1,4 +1,4 @@
-const { parseOptions } = require('../lib/parse-options');
+const { parseOptions } = require('./parse-options');
 
 /**
  * Test CSS option value collection and normalization.
