+++ lib/tidy-shorthand-property.test.js
@@ -1,5 +1,5 @@
 const postcss = require('postcss');
-const tidyShorthandProperty = require('../lib/tidy-shorthand-property');
+const tidyShorthandProperty = require('./tidy-shorthand-property');
 
 /**
  * Create a test plugin to replace shorthand properties.
