+++ lib/utils/parseAtruleParams.test.js
@@ -1,0 +1,43 @@
+const parseAtruleParams = require('./parseAtruleParams');
+
+describe('Parse media queries', () => {
+  test('min-width', () => {
+    expect(parseAtruleParams('(min-width: 100px)'))
+      .toEqual([{ minMax: 'min', value: '100px' }]);
+  });
+
+  test('max-width', () => {
+    expect(parseAtruleParams('(max-width: 200px)'))
+      .toEqual([{ minMax: 'max', value: '200px' }]);
+  });
+
+  test('min-width and max-width', () => {
+    expect(parseAtruleParams('(min-width: 100px) and (max-width: 200px)'))
+      .toEqual([{ minMax: 'min', value: '100px' }, { minMax: 'max', value: '200px' }]);
+  });
+
+  test('screen and min-width', () => {
+    expect(parseAtruleParams('screen and (min-width: 200px)'))
+      .toEqual([{ minMax: 'min', value: '200px' }]);
+  });
+
+  test('screen and max-width', () => {
+    expect(parseAtruleParams('screen and (max-width: 200px)'))
+      .toEqual([{ minMax: 'max', value: '200px' }]);
+  });
+
+  test('screen and min-width and max-width', () => {
+    expect(parseAtruleParams('screen and (min-width: 100px) and (max-width: 200px)'))
+      .toEqual([{ minMax: 'min', value: '100px' }, { minMax: 'max', value: '200px' }]);
+  });
+
+  test('screen and min-width (rem) and max-width (em)', () => {
+    expect(parseAtruleParams('screen and (min-width: 10rem) and (max-width: 40em)'))
+      .toEqual([{ minMax: 'min', value: '10rem' }, { minMax: 'max', value: '40em' }]);
+  });
+
+  test('screen and print', () => {
+    expect(parseAtruleParams('screen and print'))
+      .toEqual([]);
+  });
+});
